Add files via upload
includes add_tasks, to create more tasks for a Tproc or tgroup set of tasks with a shared queue. Called:

add_tasks ?number of tasks?

with a default of 1

This needs more testing, esp. to make sure adding tasks while a task is doing work. So, not ready for a release.

Also, when we try to <tab> expand and the task entry (on the left in send_command) is a wild-card, we would error out. Might as well use something, so we use the first task from the task menu instead. If these names are bad, well, user can just skip them; better than an error.