Update tasks-1.13.tm
Now, a working puts wrapper for windows. This handles all 3 possible arguments to puts. The problem was that package commands do "puts -nonewline channel string", with a channel that is not stdout/stderr. So, these must not be diverted to putz. But for stdout/stderr, this puts wrapper diverts (windows only) to putz. putz now handles -nonewline as well. Settings for t_debug still determine what putz does. On windows, a value of 0 will use a tk window per task, but a value of 2 goes to the console, which is only in the main thread (so uses a thread::send to get it there).

On linux (not windows), puts goes to stdout, so the code is not needed there. This puts wrapper is explained in the source code comment block at the top. 

The tk bug that caused a crash on linux is fixed, according to the ticket, and the fix should be in 8.6.12. There you can get tk windows for a putz if you set t_debug to 4/5 to force linux to use tk windows with putz. I have not tested this linux fix.