Update vw_debugging.tcl
the goto line is now cognizant of the proc or method where the goto line was. This fixes the double-click of a line number in the code window that would stop at that line number in any proc or method. The command g -N can also be used to goto a line number. It should probably have an optional 2nd arg to enter the proc/method name. A goto line in a method will also save the class name, since with multiple classes, the names can be reused. Note, the same issues might be with namespaced procs and has not been tested yet. Still a work in progress.