Add files via upload
Note, there are several places the path to the tasks module and the debugger will need to be changed before running. This will run on linux or windows. If Linux, it should run w/o crashing in 8.6.12 - earlier versions have a Tk timing/locking bug. I find that with older versions tcl crashes perhaps every 10-20 times run. So, for an example, no biggy.