Update vw_debugging.tcl
clean up some instrumentation code that was still showing in methods and comment lines. The instrument+ (or alias i) command now does the eval internally so the user code (or better yet, in the command entry in the code window) does not need them. It also returns a "" so if the eval is done anyway, it will be ignored. Now tests for doing an instrument+ or a -revert while the procedure is on the call stack (methods can't revert) to avoid problems of changing code while the proc is still running. Internally stores proc names w/o the leading :: so a match when doing a -revert should not fail.