Update vw_debugging.tcl
Deal with pathological cases better, these cause a lot of shimmering when we convert lists to text. 

New menu option checkbox for coverage, by simply not erasing the code line arrows (as we would normally to move the arrow). This only works with the new minimal updates on as well. Probably not very useful if instrumenting more than one proc/method. But can turn off bp's for all but the one. Needs manual update.