Update vw_debugging.tcl
incorporate the widget tree browser from the tasks module into the debugger for when it's used standalone. Added a if 0'd out tiny tester at the bottom. cputs in this code checks for the console, if not exist, it outputs to stdout (unix). Looks like 8.7 unix will have a console, but before that...

 a right-click on any node in the widget tree opens the tree at at that point