Update vw_debugging.tcl
new instrument menu to allow for picking a proc to instrument (not methods however). It can also revert. Two new config items to suppress many global procs that are not in the user program. Increased and parameterized the max number of frames it will work with. Set to 20 (but has to be -20 in the code).