Update vw_debugging.tcl
statements with an on-line comment (using ;# at the end of the line) if later commented out completely, would not insert the lbp+ causing the next statement to not be stepped. some extra debugging statements were used to discover this and are left in but also now commented out. additional commenting and clarification - blank lines and commented out lines are in fact instrumented. This is so the coverage tracks appear on blank and commented lines. imho, it looks better this way, otherwise there'd be a lot of "holes" in the coverage tracking. In addition, it enables the double-click/go to line to work on these lines as well.