Update tasks-1.13.tm
The two error catchers now allow a choice with a yes/no tk_messageBox, yes will exit the program, no will just suspend the task. If suspend only, it will just close the message to reduce screen clutter, while allowing one to view the pieces and send commands.

When there's an error in task startup, the task monitor and tdump can see the error message in the $t_name,error tsv variable, instead of an error that the task monitor gets (because the task is only partially set up) This is typically an error in the task initialization, such as a failed package require.