Update vw_debugging.tcl
adds <esc> to close brackets in the uplevel and command text entries. A selection of code in the code section, followed by a right click will copy the selection to the uplevel entry at the point of the insertion cursor. A button is added to switch between tabsize 4 and 8 - lots of tcl library code indents using tabs and spaces assuming a tabsize of 8. A tabsize of 4 treats all tabstops at 4 char intervals.