Update tasks-1.13.tm
place a max string size on the task_monitor's result text entry in case  of a very large task result since entries can't handle huge (e.g. megabyte) strings - now an extra argument when running, which defaults to 500 - so the compare of a new value might give a false negative for color change as it only tests the max number of chars for a difference. 
increased the range for refresh to .2 to 5 by .2 - can still edit the spinbox directly, often use .1